requirements.txt
asyncpg==0.26.0
graphene==3.1.1
gunicorn==20.1.0
starlette-graphene3==0.6.0
starlette==0.20.4
tortoise-orm==0.19.2
uvicorn==0.18.3
wsproto==1.2.0
